renderer: start splitting into functional modules
